requirements.txt
sentence-transformers
torch 
torchvision
torchaudio
accelerate
transformers
pdfplumber
pytesseract
pillow
pymupdf
pymupdf_layout
faiss-cpu
openai
tqdm
numpy
rank-bm25
python-dotenv
pandas